+++ bits/src/int_test.rs
@@ -1,7 +1,81 @@
 use core::fmt::Debug;
 
-use bits::Lsb;
-use num::Int;
+use bits::Int;
+
+#[test]
+fn lsb() {
+    let tests = [
+        (0b_0000_0000_u8, 0b0000_0000),
+        (0b_0000_0001_u8, 0b0000_0001),
+        (0b_0000_1100_u8, 0b0000_0100),
+        (0b_1001_0100_u8, 0b0000_0100),
+        (0b_1001_0000_u8, 0b0001_0000),
+    ];
+
+    for (n, want) in tests {
+        assert_eq!(n.lsb(), want);
+        assert_eq!((n as i8).lsb(), want as i8);
+    }
+}
+
+#[test]
+fn msb() {
+    let tests = [
+        (0b_0000_0000_u8, 0b_0000_0000_u8),
+        (0b_0000_0001_u8, 0b_0000_0001_u8),
+        (0b_0000_0011_u8, 0b_0000_0010_u8),
+        (0b_0000_1100_u8, 0b_0000_1000_u8),
+        (0b_1001_0100_u8, 0b_1000_0000_u8),
+        (0b_1001_0000_u8, 0b_1000_0000_u8),
+    ];
+
+    for (n, want) in tests {
+        assert_eq!(n.msb(), want);
+        assert_eq!((n as i8).msb(), want as i8);
+    }
+}
+
+#[test]
+fn uint_msb_assertion() {
+    assert_eq!(0u8.msb(), 0);
+    assert_eq!(1u8.msb(), 1);
+    assert_eq!(2u8.msb(), 2);
+    assert_eq!(3u8.msb(), 2);
+    assert_eq!(4u8.msb(), 4);
+    assert_eq!(5u8.msb(), 4);
+    assert_eq!(6u8.msb(), 4);
+    assert_eq!(7u8.msb(), 4);
+    assert_eq!(8u8.msb(), 8);
+    assert_eq!(9u8.msb(), 8);
+    assert_eq!(10u8.msb(), 8);
+    assert_eq!(15u8.msb(), 8);
+    assert_eq!(16u8.msb(), 16);
+    assert_eq!(18u8.msb(), 16);
+    assert_eq!(30u8.msb(), 16);
+    assert_eq!(33u8.msb(), 32);
+}
+
+#[test]
+fn sint_msb_assertion() {
+    assert_eq!((-1i8).msb(), -128);
+
+    assert_eq!(0i8.msb(), 0);
+    assert_eq!(1i8.msb(), 1);
+    assert_eq!(2i8.msb(), 2);
+    assert_eq!(3i8.msb(), 2);
+    assert_eq!(4i8.msb(), 4);
+    assert_eq!(5i8.msb(), 4);
+    assert_eq!(6i8.msb(), 4);
+    assert_eq!(7i8.msb(), 4);
+    assert_eq!(8i8.msb(), 8);
+    assert_eq!(9i8.msb(), 8);
+    assert_eq!(10i8.msb(), 8);
+    assert_eq!(15i8.msb(), 8);
+    assert_eq!(16i8.msb(), 16);
+    assert_eq!(18i8.msb(), 16);
+    assert_eq!(30i8.msb(), 16);
+    assert_eq!(33i8.msb(), 32);
+}
 
 trait Pdep {
     fn pdep(self, mask: Self) -> Self;
@@ -52,7 +126,7 @@ unsafe fn pdep_u64_bmi2(n: u64, mask: u64) -> u64 {
     unsafe { _pdep_u64(n, mask) }
 }
 
-fn _pdep<T: Int + num::Arith + num::BitwiseAssign + bits::Block + Lsb>(data: T, mut mask: T) -> T {
+fn _pdep<T: Int>(data: T, mut mask: T) -> T {
     let mut dest = T::ZERO;
     for i in 0..<T as bits::Block>::BITS {
         if !mask.any() {
@@ -68,7 +142,7 @@ fn _pdep<T: Int + num::Arith + num::BitwiseAssign + bits::Block + Lsb>(data: T, 
 
 fn pdep_test<T>(s: T, m: T, o: T)
 where
-    T: Int + num::Arith + num::BitwiseAssign + bits::Block + Lsb + Pdep + Debug,
+    T: Int + Pdep + Debug,
 {
     assert_eq!(s.pdep(m), o);
     #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
